+++ lib/removeWordPress.py
@@ -23,6 +23,7 @@ along with WPHardening.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
 import logging
+import shutil
 from lib.termcolor import colored
 
 
@@ -39,6 +40,20 @@ class removeWordPress():
         self.directory = os.path.abspath(directory)
         self.readme = "/readme.html"
         self.license = ["/licencia.txt", "/license.txt"]
+        self.themes = [
+            "/wp-content/themes/twentytwelve/",
+            "/wp-content/themes/twentythirteen/"
+        ]
+
+    def deleteThemes(self):
+        """
+        :return: None
+        """
+        for pathThemes in self.themes:
+            if os.path.exists(self.directory + pathThemes):
+                shutil.rmtree(self.directory + pathThemes)
+                logging.info("Delete: Theme " + pathThemes)
+                print colored('\tdelete:\ttheme ' + pathThemes, 'red')
 
     def deleteReadme(self):
         """
@@ -66,3 +81,4 @@ class removeWordPress():
         print colored('\nRemove files by defaults', 'yellow')
         self.deleteReadme()
         self.deleteLicense()
+        self.deleteThemes()