+++ lib/wpconfigWordPress.py
@@ -66,6 +66,7 @@ class wpconfigWordPress():
         self.salt = self.getSalt()
         self.setTablePrefix()
         self.setLanguage()
+        self.setMemoryLimit()
         self.setWpCron()
         self.setSslCertificate()
         self.getCompletConfig()
@@ -75,7 +76,7 @@ class wpconfigWordPress():
         :return: None
         """
         value = raw_input('\tName of the database > ')
-        if value == '':
+        if value.strip() == '':
             self.setDbName()
         else:
             self.db_name = value
@@ -85,7 +86,14 @@ class wpconfigWordPress():
         :return: None
         """
         value = raw_input('\tName of the User > ')
-        if value == '':
+        if value.lower().strip() == 'root':
+            print (
+                colored(
+                    '\n\tThe use of the root user is not recommended.', 'red'
+                )
+            )
+            self.setDbUser()
+        elif value.strip() == '':
             self.setDbUser()
         else:
             self.db_user = value
@@ -95,7 +103,7 @@ class wpconfigWordPress():
         :return: None
         """
         value = raw_input('\tPassword of the user > ')
-        if value == '':
+        if value.strip() == '':
             self.setDbPassword()
         else:
             self.db_password = value
@@ -105,7 +113,7 @@ class wpconfigWordPress():
         :return: None
         """
         value = raw_input('\tHost [localhost] > ')
-        if value == '':
+        if value.strip() == '':
             self.db_host = 'localhost'
         else:
             self.db_host = value
@@ -115,7 +123,7 @@ class wpconfigWordPress():
         :return: None
         """
         value = raw_input('\tTable prefix [wph_] > ')
-        if value == '':
+        if value.strip() == '':
             self.table_prefix = 'wph_'
         else:
             self.table_prefix = value
@@ -125,10 +133,20 @@ class wpconfigWordPress():
         :return: None
         """
         value = raw_input('\tLanguage [es_ES] > ')
-        if value == '':
+        if value.strip() == '':
             self.language = ''
         else:
             self.language = value
+
+    def setMemoryLimit(self):
+        """
+        :return: None
+        """
+        value = raw_input('\tMemory Limit [64M] > ')
+        if value.strip() == '':
+            self.memorylimit = '64M'
+        else:
+            self.memorylimit = value
 
     def setWpCron(self):
         """
@@ -273,6 +291,18 @@ require_once(ABSPATH . 'wp-settings.php');
                 'WordPress Localized Language, defaults to English'
             ) +
             'define(\'WPLANG\', \'' + self.language + '\');\n\n'
+        )
+        f.write(
+            self.getComment(
+                'WordPress Memory Limit'
+            ) +
+            'define(\'WP_MEMORY_LIMIT\', \'' + self.memorylimit + '\');\n\n'
+        )
+        f.write(
+            self.getComment(
+                'Disable reporting error'
+            ) +
+            'error_reporting(0);\n' + '@ini_set(\'display_errors\', 0);\n\n'
         )
         if self.wpcron == 'true':
             f.write(